% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpfun.R
\name{interpfun}
\alias{interpfun}
\title{Interpolation function}
\usage{
interpfun(x, y, z, method = "linear")
}
\arguments{
\item{x, y}{two numeric vectors of the same size}

\item{z}{a nummeric vector of the same size as \code{x} and \code{y}}

\item{method}{method of interpolation, either \code{"linear"} or 
\code{"sibson"}}
}
\value{
A function whose graph interpolates the points \code{((x,y), z)}.
}
\description{
Generates a function \code{f(x,y)} that interpolates the known 
  function values at some given \code{(x,y)}-coordinates.
}
\examples{
library(interpolation)
}
